+++ Tests/Mapping/Loader/YamlFileLoaderTest.php
@@ -12,8 +12,6 @@
 namespace Symfony\Component\Validator\Tests\Mapping\Loader;
 
 use PHPUnit\Framework\Attributes\DataProvider;
-use PHPUnit\Framework\Attributes\Group;
-use PHPUnit\Framework\Attributes\IgnoreDeprecations;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Validator\Constraints\All;
 use Symfony\Component\Validator\Constraints\Callback;
@@ -26,7 +24,6 @@ use Symfony\Component\Validator\Mapping\ClassMetadata;
 use Symfony\Component\Validator\Mapping\Loader\YamlFileLoader;
 use Symfony\Component\Validator\Tests\Dummy\DummyGroupProvider;
 use Symfony\Component\Validator\Tests\Fixtures\Attribute\GroupProviderDto;
-use Symfony\Component\Validator\Tests\Fixtures\CallbackClass;
 use Symfony\Component\Validator\Tests\Fixtures\ConstraintA;
 use Symfony\Component\Validator\Tests\Fixtures\ConstraintB;
 use Symfony\Component\Validator\Tests\Fixtures\ConstraintWithRequiredArgument;
@@ -131,24 +128,6 @@ class YamlFileLoaderTest extends TestCase
         $expected->addGetterConstraint('lastName', new NotNull());
         $expected->addGetterConstraint('valid', new IsTrue());
         $expected->addGetterConstraint('permissions', new IsTrue());
-
-        $this->assertEquals($expected, $metadata);
-    }
-
-    #[IgnoreDeprecations]
-    #[Group('legacy')]
-    public function testLoadClassMetadataValueOption()
-    {
-        $loader = new YamlFileLoader(__DIR__.'/constraint-mapping-value-option.yml');
-        $metadata = new ClassMetadata(Entity::class);
-
-        $loader->loadClassMetadata($metadata);
-
-        $expected = new ClassMetadata(Entity::class);
-        $expected->addConstraint(new Callback('validateMeStatic'));
-        $expected->addConstraint(new Callback([CallbackClass::class, 'callback']));
-        $expected->addPropertyConstraint('firstName', new Type(type: 'string'));
-        $expected->addPropertyConstraint('firstName', new Choice(choices: ['A', 'B']));
 
         $this->assertEquals($expected, $metadata);
     }